Modified code to deal with BeamConnect
BeamConnect is a service that allows remote access to APIs via Beam.
It uses proxying to achieve this, so I had to add the relevant
functionality.

As it turned out, I was not able to get this to work with Spring
WebClient, but it did work with Apache HTTP. In this intermediate
commit, I have included both libraries, but in the future, this
will be pared down to one again.