Added variant search, BMI, height and weight
Moved the original "variant search" functionality to HGVS, because
that describes better what it actually queries. I also found out
from the Beacon support that the best way to do this is as an
extension to the URL rather than as JSON in the query body.
So I had to come up with a way of adding these to the filter list
And not have them being turned into filters. I generalized the
filter list so that it now carries a number of data types, but
I still build it in the same way, by flattening the AST and then
converting each AST node into a separate query element in a list.
It works!

The new variant search works using Beacon requestParameters rather
than filters. So I added a request parameters type to the query
elements. This is treated in a slightly different way to filters,
because there can only be one request parameters block in the query
JSON. In this case, I merge all of them together, which fits with
the piecemeal way that the current GUI operates.
It doesn't work though. I am waiting for a reply from Beacon support
regarding this.

I also implemented searches for BMI, weight and height, using range
queries built into filters. These are documented on the Beacon site,
but I have not been able to get them to work yet. For these I also
have a pending Beacon support ticket.