Refinements for GDI publication
GDI-specific site definitions were written, which work with the Cineca
data that is supplied with the GDI starter kit (a Beacon instance
written in Python).

A new "treatment" stratifier was written, to test large stratifiers.

Timers were inserted into the query code, to help gather data for the
publication.

The Dockerfile needed to be edited to get version consistency.

The GitHub workflow needed to be updated to push to Docker Hub properly.